[AN] refactor: DomainException 패키지 구조 변경 및 Mapper 분리 #742
*   [AN] refactor: DomainException 패키지 이동 (domain -> domain.exception)
*   [AN] refactor: 예외 매핑 로직을 data 모듈의 ExceptionMapper로 분리
*   [AN] refactor: ResultDomainMapper 제거 및 NetworkResultExtensions로 대체
*   [AN] refactor: ViewModel 및 Repository의 import 경로 수정
