+++ android/app/src/main/java/com/onair/hearit/data/mapper/ExceptionMapper.kt
@@ -1,17 +1,10 @@
-package com.onair.hearit.domain
+package com.onair.hearit.data.mapper
 
 import com.onair.hearit.data.datasource.NetworkResult
+import com.onair.hearit.domain.exception.DomainException
 
-sealed class DomainException(
-    message: String,
-) : Throwable(message) {
-    data object NetworkConnection : DomainException(ERROR_CHECK_NETWORK_MESSAGE)
-
-    data object UserNotRegistered : DomainException(ERROR_UNAUTHORIZED_MESSAGE)
-}
-
-object DomainExceptionMapper {
-    fun toDomainException(failure: NetworkResult.Failure): Throwable =
+object ExceptionMapper {
+    fun mapToThrowable(failure: NetworkResult.Failure): Throwable =
         when (failure) {
             // 사용자 액션 불필요 → IllegalStateException (시스템 오류)
             is NetworkResult.Failure.BadRequest -> {
@@ -37,8 +30,6 @@ object DomainExceptionMapper {
         }
 }
 
-private const val ERROR_CHECK_NETWORK_MESSAGE = "네트워크 연결을 확인해주세요"
-private const val ERROR_UNAUTHORIZED_MESSAGE = "로그인이 필요합니다"
 private const val ERROR_UNKNOWN_MESSAGE = "알 수 없는 오류가 발생했습니다"
 private const val ERROR_SERVER_MESSAGE = "서버 오류"
 private const val ERROR_BAD_REQUEST_MESSAGE = "잘못된 요청:"
