+++ android/app/src/main/java/com/onair/hearit/data/repository/PlayingHistoryRepositoryImpl.kt
@@ -3,6 +3,8 @@ package com.onair.hearit.data.repository
 import com.onair.hearit.data.datasource.remote.PlayingHistoryRemoteDataSource
 import com.onair.hearit.data.dto.PlayingHistoryRequest
 import com.onair.hearit.data.mapper.toDomain
+import com.onair.hearit.data.toDomainResult
+import com.onair.hearit.data.toDomainResultList
 import com.onair.hearit.domain.model.PlayingHistoryHearit
 import com.onair.hearit.domain.repository.PlayingHistoryRepository
