+++ android/app/src/main/java/com/onair/hearit/presentation/splash/SplashViewModel.kt
@@ -9,8 +9,8 @@ import com.onair.hearit.R
 import com.onair.hearit.analytics.AnalyticsEventNames
 import com.onair.hearit.di.AnalyticsProvider
 import com.onair.hearit.di.TokenInterceptorProvider
-import com.onair.hearit.domain.DomainException.NetworkConnection
-import com.onair.hearit.domain.DomainException.UserNotRegistered
+import com.onair.hearit.domain.exception.DomainException.NetworkConnection
+import com.onair.hearit.domain.exception.DomainException.UserNotRegistered
 import com.onair.hearit.domain.repository.AuthRepository
 import com.onair.hearit.presentation.SingleLiveData
 import kotlinx.coroutines.delay
