+++ android/app/src/test/java/com/onair/hearit/repository/UserRepositoryTest.kt
@@ -5,7 +5,7 @@ import com.onair.hearit.data.datasource.local.UserLocalDataSource
 import com.onair.hearit.data.datasource.remote.UserRemoteDataSource
 import com.onair.hearit.data.dto.UserInfoResponse
 import com.onair.hearit.data.mapper.toDomain
-import com.onair.hearit.data.repository.toDomainResult
+import com.onair.hearit.data.toDomainResult
 import com.onair.hearit.domain.model.UserInfo
 import com.onair.hearit.domain.repository.UserRepository
 import io.mockk.coEvery
